Pull class into own file
